# openapi.yaml
openapi: 3.0.0
info:
  title: Моковое API Пиццерии
  description: API для работы с пиццами в моковой пиццерии
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /api/login:
    post:
      summary: Авторизация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успешная авторизация. Возвращает токен доступа.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
  /api/pizzas:
    get:
      summary: Получение списка пицц
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Успешный запрос. Возвращает список пицц.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pizza'
    post:
      summary: Создание новой пиццы
      security:
        - BasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pizza'
      responses:
        '201':
          description: Пицца успешно создана. Возвращает созданную пиццу.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
  /api/pizzas/{id}:
    get:
      summary: Получение информации о пицце по ID
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешный запрос. Возвращает информацию о пицце.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
        '404':
          description: Пицца с указанным ID не найдена.
    put:
      summary: Обновление информации о пицце по ID
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - in: requestBody
          required: true
          description: Обновленная информация о пицце
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
      responses:
        '200':
          description: Успешное обновление информации о пицце. Возвращает обновленную пиццу.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
    delete:
      summary: Удаление пиццы по ID
      security:
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Пицца успешно удалена.
        '404':
          description: Пицца с указанным ID не найдена.
components:
  schemas:
    Pizza:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
      required:
        - name
        - description
securitySchemes:
  BasicAuth:
    type: http
    scheme: basic